+++ resources/views/components/tools/toolbar.blade.php
@@ -206,7 +206,7 @@
             @endif
 
             @if ($component->columnSelectIsEnabled())
-                <div class="mb-4 w-full md:w-auto md:mb-0 md:ml-2">
+                <div class="@if ($component->getColumnSelectIsHiddenOnMobile()) hidden sm:block @elseif ($component->getColumnSelectIsHiddenOnTablet()) hidden md:block @endif mb-4 w-full md:w-auto md:mb-0 md:ml-2">
                     <div
                         x-data="{ open: false }"
                         @keydown.window.escape="open = false"
@@ -494,7 +494,7 @@
             @endif
 
             @if ($component->columnSelectIsEnabled())
-                <div class="mb-3 mb-md-0 pl-0 pl-md-2">
+                <div class="@if ($component->getColumnSelectIsHiddenOnMobile()) d-none d-sm-block @elseif ($component->getColumnSelectIsHiddenOnTablet()) d-none d-md-block @endif mb-3 mb-md-0 pl-0 pl-md-2">
                     <div
                         x-data="{ open: false }"
                         x-on:keydown.escape.stop="open = false"
@@ -756,7 +756,7 @@
             @endif
 
             @if ($component->columnSelectIsEnabled())
-                <div class="mb-3 mb-md-0 md-0 ms-md-2">
+                <div class="@if ($component->getColumnSelectIsHiddenOnMobile()) d-none d-sm-block @elseif ($component->getColumnSelectIsHiddenOnTablet()) d-none d-md-block @endif mb-3 mb-md-0 md-0 ms-md-2">
                     <div
                         x-data="{ open: false }"
                         x-on:keydown.escape.stop="open = false"