+++ resources/views/datatable.blade.php
@@ -1,4 +1,8 @@
 <x-livewire-tables::wrapper :component="$this">
+    @if ($this->hasConfigurableAreaFor('before-tools'))
+        @include($this->getConfigurableAreaFor('before-tools'), $this->getParametersForConfigurableArea('before-tools'))
+    @endif
+
     <x-livewire-tables::tools>
         <x-livewire-tables::tools.sorting-pills />
         <x-livewire-tables::tools.filter-pills />