+++ src/Traits/WithColumnSelect.php
@@ -2,6 +2,7 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits;
 
+use Rappasoft\LaravelLivewireTables\Events\ColumnsSelected;
 use Rappasoft\LaravelLivewireTables\Traits\Configuration\ColumnSelectConfiguration;
 use Rappasoft\LaravelLivewireTables\Traits\Helpers\ColumnSelectHelpers;
 
@@ -13,6 +14,8 @@ trait WithColumnSelect
     public array $selectedColumns = [];
     protected bool $columnSelectStatus = true;
     protected bool $rememberColumnSelectionStatus = true;
+    protected bool $columnSelectHiddenOnMobile = false;
+    protected bool $columnSelectHiddenOnTablet = false;
 
     public function setupColumnSelect(): void
     {
@@ -58,6 +61,7 @@ trait WithColumnSelect
     {
         $this->{$this->tableName}['columns'] = [];
         $this->forgetColumnSelectSession();
+        event(new ColumnsSelected($this->getColumnSelectSessionKey(), $this->selectedColumns));
     }
 
     public function deselectAllColumns()
@@ -65,6 +69,7 @@ trait WithColumnSelect
         $this->{$this->tableName}['columns'] = [];
         $this->selectedColumns = [];
         session([$this->getColumnSelectSessionKey() => []]);
+        event(new ColumnsSelected($this->getColumnSelectSessionKey(), $this->selectedColumns));
     }
 
     public function updatedSelectedColumns(): void
@@ -75,6 +80,7 @@ trait WithColumnSelect
         } else {
             $this->{$this->tableName}['columns'] = $this->selectedColumns;
             session([$this->getColumnSelectSessionKey() => $this->{$this->tableName}['columns']]);
+            event(new ColumnsSelected($this->getColumnSelectSessionKey(), $this->selectedColumns));
         }
     }